# sql: improve setseed builtin