--- a/pkg/sql/distsql/BUILD.bazel
+++ b/pkg/sql/distsql/BUILD.bazel
@@ -28,7 +28,6 @@ go_library(
         "//pkg/util/log",
         "//pkg/util/mon",
         "//pkg/util/pprofutil",
-        "//pkg/util/randutil",
         "//pkg/util/timeutil",
         "//pkg/util/tochar",
         "//pkg/util/tracing",