--- a/pkg/sql/sem/eval/BUILD.bazel
+++ b/pkg/sql/sem/eval/BUILD.bazel
@@ -85,6 +85,7 @@ go_library(
         "//pkg/util/hlc",
         "//pkg/util/json",
         "//pkg/util/mon",
+        "//pkg/util/randutil",
         "//pkg/util/rangedesc",
         "//pkg/util/ring",
         "//pkg/util/timeofday",