This commit improves https://github.com/cockroachdb/cockroach/commit/b51da9ebe2d1c2e4b7ecf810edb5c88cdc991adb in
a couple of ways.

First, it makes the allocation of the RNG lazy, whenever either random()
or setseed() is called for the first time.

Here is the impact on `flowinfra/BenchmarkInfrastructure`:
```
name                         old time/op    new time/op    delta
Infrastructure/n1/r1-24        37.0µs ± 1%    23.0µs ± 2%  -37.74%  (p=0.000 n=10+10)
Infrastructure/n1/r100-24      87.7µs ± 1%    75.7µs ± 2%  -13.76%  (p=0.000 n=8+10)
Infrastructure/n1/r10000-24    5.03ms ± 2%    5.06ms ± 1%     ~     (p=0.165 n=10+10)
Infrastructure/n3/r1-24         749µs ± 4%     714µs ± 4%   -4.57%  (p=0.000 n=10+10)
Infrastructure/n3/r100-24      1.01ms ± 3%    0.96ms ± 4%   -5.09%  (p=0.000 n=10+10)
Infrastructure/n3/r10000-24    19.4ms ± 1%    19.6ms ± 1%     ~     (p=0.247 n=10+10)

name                         old alloc/op   new alloc/op   delta
Infrastructure/n1/r1-24        28.2kB ± 2%    22.3kB ± 1%  -20.77%  (p=0.000 n=9+8)
Infrastructure/n1/r100-24      74.2kB ± 0%    68.4kB ± 0%   -7.81%  (p=0.000 n=8+8)
Infrastructure/n1/r10000-24    4.58MB ± 0%    4.57MB ± 0%     ~     (p=0.161 n=8+8)
Infrastructure/n3/r1-24         168kB ± 1%     153kB ± 9%   -8.85%  (p=0.000 n=8+9)
Infrastructure/n3/r100-24       401kB ± 0%     383kB ± 1%   -4.47%  (p=0.000 n=8+9)
Infrastructure/n3/r10000-24    21.9MB ± 0%    22.0MB ± 2%     ~     (p=0.481 n=8+9)

name                         old allocs/op  new allocs/op  delta
Infrastructure/n1/r1-24           122 ± 3%       116 ± 1%   -4.64%  (p=0.000 n=9+8)
Infrastructure/n1/r100-24         236 ± 0%       230 ± 0%   -2.28%  (p=0.000 n=8+8)
Infrastructure/n1/r10000-24     10.4k ± 0%     10.4k ± 0%     ~     (p=0.151 n=8+8)
Infrastructure/n3/r1-24         1.52k ± 1%     1.50k ± 2%   -1.47%  (p=0.005 n=8+8)
Infrastructure/n3/r100-24       2.20k ± 0%     2.17k ± 1%   -1.16%  (p=0.000 n=8+9)
Infrastructure/n3/r10000-24     54.0k ± 0%     54.6k ± 5%     ~     (p=0.743 n=8+9)
```

Second, it fixes the logic so that all consequent statements on the same
session do use the right seed - previously, subqueries would use
a separate RNG created in `initEvalCtx`.

This refactor also allows us to only set the RNG field on the eval
context on the main code path.

Epic: None

Release note: None