@yuzefovich 2024-04-02
I think the only concern with this change is whether we need to make access to the RNG atomic. `rand.Rand` object itself is not concurrency safe IIRC, so before this change calls on `evalCtx.RNG` already could be racy; this change makes also the instantiation of `.rng` racy.